Everything now compiles with reasonably few warnings, except tedomari.
Let there be rejoicing in the streets.

I changed my editor to leave spaces, but there may be some files that I edited beforehand, and didn't change back.

I probably broke the gcc build somewhere here, for which apologies. 

git-svn-id: file:///var/lib/svn/sakusen/cpp/trunk@496 6f43db7c-3df8-0310-95f7-eae406835aab

+++ libsakusen/client/updatedunit.h
@@ -17,7 +17,8 @@ class UpdatedUnit : public CompleteUnit {
   public:
     UpdatedUnit(const CompleteUnit& copy) :
       CompleteUnit(copy),
-      orders(copy.getStatus().getWeaponsStatus().size()),
+	  orders(static_cast<uint16>(copy.getStatus().getWeaponsStatus().size())),
+	  
       altered(false)
     {}
     virtual ~UpdatedUnit() {}
